change to thread display of messages
